+++ austrakka/utils/enums/api.py
@@ -1,0 +1,9 @@
+RESPONSE_TYPE_SUCCESS = 'Success'
+RESPONSE_TYPE_WARNING = 'Warning'
+RESPONSE_TYPE_ERROR = 'Error'
+RESPONSE_TYPE = 'ResponseType'
+RESPONSE_MESSAGE = 'ResponseMessage'
+RESPONSE_ROW_NUMBER = 'RowNumber'
+
+RESPONSE_DATA = 'data'
+RESPONSE_MESSAGES = 'messages'